+++ DatsartSpace/API/Models/Shot.cs
@@ -2,8 +2,8 @@
 
 public class Shot
 {
-    public int AngleHorizontal { get; set; }
-    public int AngleVertical { get; set; }
+    public float AngleHorizontal { get; set; }
+    public float AngleVertical { get; set; }
     public object ColorPot { get; set; }
-    public int Power { get; set; }
+    public float Power { get; set; }
 }
